requirements.txt
pandas>=2.2.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
pyinstaller>=6.0.0

